// client/src/types/github/githubTree.ts
export enum TreeNodeType {
	file = "file",
	folder = "folder",
}

export interface TreeNode {
	name: string;
	path: string;
	type: TreeNodeType;
	children: TreeNode[];
	isExpanded: boolean;
}

export enum GithubTreeItemType {
	blob = "blob",
	tree = "tree",
	commit = "commit",
}

export interface GithubTreeItem {
	path: string;
	type: GithubTreeItemType;
	mode: string;
	sha: string;
	url: string;
	size?: number;
}